Import Docker ipv6 network fixture.
